Studious Seal Premium/requirements.txt
discord.py==2.3.2
python-telegram-bot==20.7
google-genai==0.3.0
pydantic==2.6.3
python-dotenv==1.0.1
